my-product-agent/requirements.txt
fastapi
uvicorn
redis
pydantic-settings
google-generativeai